`SimEngine` re-write and minor changes
The re-write to `SimEngine` comes from a general shift of direction in the focus of the project. Rather than aiming to  simulate around known economic theory, `macrosim` now takes a significantly more exploratory approach to economic model discovery. In conjunction with `SeriesAccessort` and `EqSearch`, the library is equipped to form any kinds of economic datasets to model any arbitrary metric. Note that in a future update `EqSearch` planned to switch to a significantly tuned down instance of `PySRRegressor` in favor of interpretability. Current equation outputs are significantly complex and although they are accurate as estimators, the features often don't interact in a way that conventional economics would define. (For example, a higher labor participation rate can result in lower GDP figures past 70% in testing) As of this commit, `SimEngine` accepts a sympy equation as an argument and model the simulation through growth functions defined individually for each parameter the equation takes. This format allows for much more freedom for working with expressions derived from `macrosim.EqSearch` .  Switching to this format requires a certain level of interpretability to allow the exploration of conventionally plausible growth functions. A new class, `GrowthPatternDetector` will be implemented in future commits to fit well-known growth patterns to data and derive the most plausible pattern as a default for all simulation applications.