`GrowthPatternDetector` initial implementation
The first iteration of `GrowthPatternDetector` is implemented, though it needs significant re-touching. Current line fitting behavior is strugging with assessing a general growth pattern for time series. Most likely solution will be to move away from hard coded functions with coefficients to fit, and moving to time-series analysis. As a note, the current fit outputs are much worse than using the average rate of change for each column in the dataset.