`EqSearch` parameter update and `SeriesAccessor` fill method additions
Added 3 common NaN handling strategies  as built-in methods to the fill method. `EqSearch now accepts `kwargs` to customize some parameters of `PySRRegressor` that were previously inaccessible to users. Equation constraints are now available for customization and trig functions are now removed from the default set of unary operations in favour of simplicity.